Allow negative coinbase profit in `Order::Tx` and `Order::Bundle` (#233)
When disallowed for all `Order` varients, the coinbase account has no
way to spend funds when rbuilder is building all blocks for a chain.

Instead of using saturating_sub, I also tried changing `coinbase_profit`
from a `U256` to an `I256`, but ran into issues where a `U256` is
required to calculate stuff like `mev_gas_price` in later logic. I'd
prefer that approach if anyone knows if it is feasible.